[Improvement][hudi-spark-connector] Moved Config class to package hudi based on review comments
